- factored out heap redirection code; - fixed missing mutex in StaticBufferHeap; - moved memalign FIXME to StaticBufferHeap, where it can more easily be   fixed;
